Crate sidechain_runtime

Re-exports

pub use pallet_session_validator_management;

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

BlockLength

BlockWeights

ConstBool

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstU128

Const getter for a basic type.

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FeeMultiplier

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

LogBeneficiaries

MaxTransactions

MaxValidators

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

StorageInfo

Metadata about storage from the runtime.

TokenConversionRate

Version

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

AVERAGE_ON_INITIALIZE_RATIO

DAYS

EXISTENTIAL_DEPOSIT

Existential deposit.

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time.

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

Balance

Balance of an account.

Balances

BalancesConfig

BeneficiaryId

Block

Block type as expected by this runtime.

BlockNumber

An index to a block.

BlockRewards

CrossChainPublic

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Nonce

Index of a transaction in the chain.

Session

SessionCommitteeManagement

SessionCommitteeManagementConfig

SessionConfig

Sidechain

SidechainConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.